[Feature][Lex] Created Carousel Component
