+++ src/components/modules/Product/index.js
@@ -1,3 +1,2 @@
 export { default as ProductDetails } from "./ProductDetails";
 export { default as ProductOrder } from "./ProductOrder";
-export { default as ProductPreview } from "./ProductPreview";
